feat: remove network-filter model and move it to network-list model
