--- src/renderer/features/networks/NetworkFilter/index.ts
@@ -1,1 +1,0 @@
-export { filterModel } from './model/network-filter';
