+++ src/renderer/features/networks/NetworkList/model/network-list.ts
@@ -1,12 +1,37 @@
-import { Store, combine } from 'effector';
+import { combine, createEvent, createStore, sample, Store } from 'effector';
 
 import { ExtendedChain, networkModel, networkUtils } from '@/src/renderer/entities/network';
-import { filterModel } from '../../NetworkFilter';
 import { getExtendedChain } from '../networks-list-utils';
+import { includes } from '@shared/lib/utils';
+
+const componentMounted = createEvent();
+
+const $filterQuery = createStore<string>('');
+const queryChanged = createEvent<string>();
+const queryReset = createEvent();
+
+$filterQuery.on(queryChanged, (_, query) => query).reset(queryReset);
+
+sample({
+  clock: componentMounted,
+  target: queryReset,
+});
+
+const $networksFiltered = combine(
+  {
+    chains: networkModel.$chains,
+    query: $filterQuery,
+  },
+  ({ chains, query }) => {
+    if (!query) return Object.values(chains);
+
+    return Object.values(chains).filter((c) => includes(c.name, query));
+  },
+);
 
 const $activeChainsSorted: Store<ExtendedChain[]> = combine(
   {
-    filteredNetworks: filterModel.$networksFiltered,
+    filteredNetworks: $networksFiltered,
     connectionStatuses: networkModel.$connectionStatuses,
     connections: networkModel.$connections,
   },
@@ -19,7 +44,7 @@ const $activeChainsSorted: Store<ExtendedChain[]> = combine(
 
 const $inactiveChainsSorted: Store<ExtendedChain[]> = combine(
   {
-    filteredNetworks: filterModel.$networksFiltered,
+    filteredNetworks: $networksFiltered,
     connectionStatuses: networkModel.$connectionStatuses,
     connections: networkModel.$connections,
   },
@@ -30,4 +55,13 @@ const $inactiveChainsSorted: Store<ExtendedChain[]> = combine(
   },
 );
 
-export const networkListModel = { $activeChainsSorted, $inactiveChainsSorted };
+export const networkListModel = {
+  $activeChainsSorted,
+  $inactiveChainsSorted,
+  $networksFiltered,
+  $filterQuery,
+  events: {
+    componentMounted,
+    queryChanged,
+  },
+};
