+++ src/renderer/features/networks/index.ts
@@ -1,2 +1,1 @@
-export { filterModel } from './NetworkFilter';
 export { networkListModel } from './NetworkList';
